requirements.txt
toml
streamlit
requests
chromadb
python-dotenv
sentence-transformers
tqdm